create edit and delete dialog componnents
