+++ src/components/screens/AdminPage/Dialogs/DeleteUserDialog.jsx
@@ -1,0 +1,32 @@
+import React from 'react';
+import {
+  Dialog,
+  DialogTitle,
+  DialogContent,
+  DialogActions,
+  Button,
+  Typography
+} from '@mui/material';
+
+const DeleteUserDialog = ({ open, onClose, onDelete }) => {
+  return (
+    <Dialog open={open} onClose={onClose} maxWidth="sm" fullWidth>
+      <DialogTitle>Delete User</DialogTitle>
+      <DialogContent>
+        <Typography>
+          {"Are you sure you want to delete this user?"}
+        </Typography>
+      </DialogContent>
+      <DialogActions>
+        <Button onClick={onClose} color="secondary">
+          Cancel
+        </Button>
+        <Button onClick={onDelete} variant="contained" color="error">
+          Delete
+        </Button>
+      </DialogActions>
+    </Dialog>
+  );
+};
+
+export default DeleteUserDialog;
